package.json - added expo media library for permissions to file system
App.js - added permission in pick multiple documents as well as changed the options for getDocumentAsync to have copyToCacheDirectory set to false to avoid a temporary file and use the permanent file locaion as the uri stored in the asynch storage

AudioHandler.js - found a bug that when clicking the multiple documents button, that function creates sound objects and replaces the audiohandler's songlist everytime which orpahns the curSong because I also change the curSong as well, so i just cleaned up the orphaning of curSong. Although there might be a more efficient solution by not recreating each sound object every time i click that button, will take some thinking because idk how to simply add a song obj to the list without recreating the whole thing. not a priority atm tho but a thought to keep in mind

screens - added new files which will be the ui for each screen on the app and the next focus is working on the song screen specifically